+++ tx.go
@@ -13,7 +13,6 @@ const (
 // Types of transaction statuses
 const (
 	StatusCompleted = "completed"
-	StatusPending   = "pending"
 	StatusFailed    = "failed"
 )
